Refs #3709 XML log tests no longer test behavior, only configuration parsed.
