+++ include/fastrtps/fastrtps_all.h
@@ -44,7 +44,6 @@
 #include "TopicDataType.h"
 
 #include "utils/IPFinder.h"
-#include "log/Log.h"
 #include "utils/eClock.h"
 #include "utils/TimeConversion.h"
